% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Extract_Votes.R
\name{Extract_Votes}
\alias{Extract_Votes}
\title{For all cases, the proportion of OOB votes that each node recieved in
each local randomForest classifer.}
\usage{
Extract_Votes(Hie_RF, Train_Predict = TRUE, New_Data = NULL,
  New_Data_Case_ID = 1, New_Data_Exp_Var = c(2:ncol(New_Data)),
  Bind_Train_New = FALSE, ...)
}
\arguments{
\item{Hie_RF}{Object of class \code{"Hier.Random.Forest"} - the output of
Run_HRF.}

\item{Train_Predict}{Logical, if TRUE, returns for each case in the
training data the proportion of OOB votes that each class received in each
local randomForesst classifer. If FALSE, only the votes for New_Data are
returned.}

\item{New_Data}{Optional data frame containing additional cases
that were note a part of the original training set, for which the proportion
of votes should be extracted.}

\item{New_Data_Case_ID}{Integer, specifying the column number with the
Case_ID in the New_Data data frame. The Case ID values should be unique
and different from those in the training data.}

\item{New_Data_Exp_Var}{Vector of integers, specifying the columns of
New_Data that contains the same set of explanatory variables as used in the
training of Hie_RF. Before running the Run_HRF, we recommend using the
Join_Levels function on each categorical variables to ensure the extraction
of votes for New_Data.}

\item{Bind_Train_New}{Logical, if TRUE the cases in the training set and
New_Data will be binded to one output data frame (along with the two
seperate ones).}

\item{...}{Optional parameters to be passed to the low level
  functions.}
}
\value{
a list consisting of up to three of the following data frames:
\item{Prop_Vote_Train}{The proportion of OOB votes that each case from the
  training dataset received in each local classifer.}
\item{Prop_Vote_New}{The proportion of OOB votes that each case from the
  New_Data dataset received in each local classifer.}
\item{Prop_Vote_Full}{Bind of Prop_Vote_Train and Prop_Vote_New if
  Bind_Train_New is TRUE.}
}
\description{
The funtion takes as input an object of class Hier.Random.Forest. For each
case in the original training data or in New_Data, the function extracts the
proportion of OOB votes for each node (internal and external) in each local
classifier.
}
\details{
For the training data, only OOB votes are used in each local
classifer.
}
\author{
Yoni Gavish <gavishyoni@gmail.com>
}

