% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Hie_F_Measure.R
\name{Hie_F_Measure}
\alias{Hie_F_Measure}
\title{Hierarchical precision, hierarchical recall and hierarchical F measure}
\usage{
Hie_F_Measure(Conf_Matr, Unique_Path, Beta_H_F = 1, By_Node = FALSE, ...)
}
\arguments{
\item{Conf_Matr}{Object of class confusionMatrix, as generated by
  confusionMatrix function of the 'caret' package.}

\item{Unique_Path}{Data frame, the Unique_Path data frame from Run_HRF.
  Can be found in an object of class Hier.Random.Forest at
  $Hier_Struc$Unique_Path}

\item{Beta_H_F}{Numeric in the range Beta_H_F>=0. Controls the
  relative weight of Hierarchical precision and hierarchical recall in the
  estimation of hierarchical F measure values close to 0 will give more
  weight to Hierarchical precision. Equal weights are for Beta_H_F=1, values
  above 1 will give higher weight to hierarchical recall. See details in:
  Kiritchenko, S., S. Matwin, and F. Famili. 2005. Functional annotation of
  genes using hierarchical text categorization. In: Proc. of the ACL
  Workshop on Linking Biological Literature, Ontologies and Databases:
  Mining Biological Semantics.}

\item{By_Node}{Logical, if TRUE, the hierachical precision, recall
  and F measure are estimated for each class as well.}

\item{Optional}{parameters to be passed to the low level funtions}
}
\description{
Hierarchical precision, hierarchical recall and hierarchical F measure
}

