% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Hie_Importance.R
\name{Hie_Importance}
\alias{Hie_Importance}
\title{The importance value of each explanatory variable in each local classifer}
\usage{
Hie_Importance(hie.RF, format.out = c("4.col.out"), scale.imp = FALSE, ...)
}
\arguments{
\item{hie.RF}{Object of class \code{"Hier.Random.Forest"} - the output of \code{Run_HRF}.}

\item{format.out}{The format of the output. can be either \code{"4.col.out"} (default), \code{"table.out"} or both. See Value for details.}

\item{scale.imp}{Logical, if \code{TRUE}, variable importance is scaled by the standard error. Default is \code{FALSE}.
See \code{\link{importance}} in the \code{randomForest} package for details.}

\item{...}{Optional parameters to be passed to the low level functions.}
}
\value{
\item{imp.val}{A data frame or list containing the importance of each explanatory varaible in each local classifier.
 \describe{
 \item{\strong{1.}   If \code{format.out = "4.col.out"}}{Data frame with four columns:}
     {\tabular{lll}{
      \code{"Classifier_ID"} \tab \tab The name of the local classifer. \cr
      \code{"Par_Level"}     \tab \tab The name of the parent node in the local classifer. \cr
      \code{"Expl_Var"}      \tab \tab The name of theexplanatory vatiable. \cr
      \code{"Mean_Dec_Accu"} \tab \tab the mean decrease in accruacy.  \cr
       }}
 \item{\strong{2.}   If \code{format.out = "table.out"}}{Data frame, with the first column containing the name of the local classifer and the rest of the columns the names and variable importance of each explantory variable.}
 \item{\strong{3.}   If \code{format.out = c("4.col.out","table.out")}}{A list containing the two data frames, under \code{"var.imp.4.col"} and \code{"var.imp.table"}, respectively.}
   }}
}
\description{
This function extracts the importance value of each variable for each local
classifier. Requires setting \code{importance = TRUE} when running \code{\link{Run_HRF}}.
Relies on the function \code{\link{importance}} in the \code{randomForest} package.
}
\details{
Returns the mean decrease in accuracy (\code{type=1} in \code{randomForest::importance}).
}
\author{
Yoni Gavish <gavishyoni@gmail.com>
}

