% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HRF_Performance.R
\name{HRF_Performance}
\alias{HRF_Performance}
\title{Mutltiple flat and hierarchical performance measures for crisp classifcation.}
\usage{
HRF_Performance(Hie_RF, Per_Index = c("Flat_Measures", "Hie_F_Measure"),
  Crisp_Rule = c("Multiplicative_Majority", "Multiplicative_Permutation",
  "Setpwise_Majority"), Perm_Num = 500, By_Node = TRUE,
  Div_Logical = TRUE, Div_Print = 25, Beta_H_F = 1, ...)
}
\arguments{
\item{Hie_RF}{Object of class Hier.Random.Forest - the output of
Run_HRF.}

\item{Per_Index}{The performance and accuracy indices to compute.See
details below.}

\item{Crisp_Rule}{The method of translating proportion of votes to a
crisp category. See details below.}

\item{Perm_Num}{Integer,number of random permutations for each case
if 'Multiplicative_Permutation' is applied.}

\item{By_Node}{Logical, if TRUE performances indices will be
estimated for each terminal node as well as for the overall confusion
matrix.}

\item{Div_Logical}{Logical, if TRUE progress when
'Multiplicative_Permutation' is applied will be printed every Div_Print
permutations}

\item{Div_Print}{Ser above}

\item{Beta_H_F}{Numeric in the range Beta_H_F>=0. Controls weights
in the hierarchical F measure index. See Hie_F_Measure for details.}

\item{...}{Optional parameters to be passed to the low level
functions.}
}
\description{
Mutltiple flat and hierarchical performance measures for crisp classifcation.
}
\author{
Yoni Gavish <gavishyoni@gmail.com>
}

