% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HRF_Performance_New_Data.R
\name{HRF_Performance_New_Data}
\alias{HRF_Performance_New_Data}
\title{predict and asses performance of New_Data, for which the 'true' class is
known.}
\usage{
HRF_Performance_New_Data(Hie_RF, New_Data, New_Data_Case_ID = 1,
  New_Data_Exp_Var = NA, New_Data_Hie,
  Crisp_Rule = c("Multiplicative_Majority", "Multiplicative_Permutation",
  "Setpwise_Majority"), Perm_Num = 500, Div_Logical = TRUE,
  Div_Print = 25, Per_Index = c("Flat_Measures", "Hie_F_Measure"),
  By_Node = TRUE, Beta_H_F = 1, ...)
}
\arguments{
\item{Hie_RF}{Object of class Hier.Random.Forest - the output of
Run_HRF.}

\item{New_Data}{The data frame containing additional cases that
were note a part of the original training.}

\item{New_Data_Case_ID}{Integer, specifying the column number with the
Case_ID in the New_Data data frame. The Case ID values should be unique and
different from those in the training data.}

\item{New_Data_Exp_Var}{Vector of integers, specifying the columns of
New_Data that contains the same set of explanatory variables as used in the
training of Hie_RF. Default is all column except New_Data_Case_ID and
New_Data_Hie.}

\item{New_Data_Hie}{Vector of character or integers, containing the
names or column numbers of the hierarchical levels in New_Data. Order of
columns should be from the tree root to the terminal nodes. If a single
column is provided, it should contain only terminal nodes.}

\item{Crisp_Rule}{The method of translating proportion of votes to a
crisp category. See details in HRF_Performance.}

\item{Perm_Num}{Integer, number of random permutations for each
case if 'Multiplicative_Permutation' is applied. See details in
HRF_Performance.}

\item{Div_Logical}{Logical, if TRUE progress when
'Multiplicative_Permutation' is applied will be printed every Div_Print
permutations}

\item{Per_Index}{The performance and accuracy indices to compute. See
details in HRF_Performance.}

\item{By_Node}{Logical, if TRUE performances indices will be
estimated for each terminal node as well as for the overall confusion
matrix.}

\item{Beta_H_F}{Numeric in the range Beta_H_F>=0. Controls weights
in the hierarchical F measure index. See Hie_F_Measure for details.}

\item{...}{Optional parameters to be passed to low level
functions.}
}
\description{
predict and asses performance of New_Data, for which the 'true' class is
known.
}
\author{
Yoni Gavish <gavishyoni@gmail.com>
}

