% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ImportanceHie.R
\name{ImportanceHie}
\alias{ImportanceHie}
\title{The importance value of each explanatory variable in each local classifier}
\usage{
ImportanceHie(hie.RF, format.out = c("col.4.out"), scale.imp = FALSE, ...)
}
\arguments{
\item{hie.RF}{Object of class \code{"HRF"} - the output of \code{RunHRF}.}

\item{format.out}{The format of the output. Can be either \code{"col.4.out"}
  (default), \code{"table.out"} or both. See Value for details.}

\item{scale.imp}{Logical, if \code{TRUE}, variable importance is scaled by
  the standard error. Default is \code{FALSE}.

See \code{\link{importance}} in the \code{randomForest} package for details.}

\item{\dots}{Optional parameters to be passed to low level functions.}
}
\value{
\item{imp.val}{A data frame or list containing the importance of each
explanatory variable in each local classifier. Output structure depends on
the settings of \code{format.out}.}

 \describe{
 \item{\strong{1.}   If \code{format.out = "col.4.out"}}{Data frame with four
 columns:

     \tabular{llllllllll}{
     \tab \tab \tab \tab \tab \tab \tab \code{"classifier.ID"} \tab \tab The
     name of the local classifier. \cr

     \tab \tab \tab \tab \tab \tab \tab \code{"par.level"}     \tab \tab The
     name of the parent node in the local classifier. \cr

     \tab \tab \tab \tab \tab \tab \tab \code{"expl.var"}      \tab \tab The
     name of the explanatory variable. \cr

     \tab \tab \tab \tab \tab \tab \tab \code{"mean.dec.accu"} \tab \tab the
     mean decrease in accuracy.  \cr }}

 \item{\strong{2.}   If \code{format.out = "table.out"}}{Data frame, with the
 first column containing the name of the local classifier and the rest of the
 columns the names and variable importance of each explanatory variable.}

 \item{\strong{3.}   If \code{format.out = c("col.4.out","table.out")}}{A
 list containing the two data frames, under \code{"var.imp.4.col"} and
 \code{"var.imp.table"}, respectively.}}
}
\description{
This function extracts the importance value of each variable for each local
classifier. Requires setting \code{importance = TRUE} when running
\code{\link{RunHRF}}. Relies on the function \code{\link{importance}} in the
\code{randomForest} package.
}
\details{
Returns the mean decrease in accuracy (\code{type=1} in
  \code{randomForest::importance}).
}
\examples{
data(OliveOilHie)
hie.RF.OO <- RunHRF(train.data       = OliveOilHie,
                   case.ID           = "case.ID",
                   hie.levels        = c(2:4),
                   mtry              = "tuneRF2",
                   internal.end.path = TRUE)

impor.hie.RF.OO <- ImportanceHie(hie.RF     = hie.RF.OO,
                                 format.out = c("col.4.out"))

plot(hie.RF.OO)
PlotImportanceHie(input.data = impor.hie.RF.OO,
                  X.data     = 2,
                  Y.data     = 3,
                  imp.data   = 4,
                  plot.type  = "Tile",
                  X.Title = c("Parent node Name"),
                  Y.Title = c("Explanatory variable"),
                  imp.title = c("Mean \\n Decrease \\n in \\n Accuracy"))

# table format
impor.hie.RF.OO <- ImportanceHie(hie.RF     = hie.RF.OO,
                                 format.out = c("table.out"))

# both output formats
impor.hie.RF.OO <- ImportanceHie(hie.RF     = hie.RF.OO,
                                 format.out = c("col.4.out","table.out"))

import.col.4.for <- impor.hie.RF.OO$imp.var.4.col
import.table.for <- impor.hie.RF.OO$imp.val.table
}
\author{
Yoni Gavish <gavishyoni@gmail.com>
}
\seealso{
\code{\link{PlotImportanceHie}} for a \code{ggplo2} based plotting
   of hierarchical importance values.
}

