% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JoinLevels.R
\name{JoinLevels}
\alias{JoinLevels}
\title{Combines the factor levels of two input vectors}
\usage{
JoinLevels(vector.1, vector.2, ...)
}
\arguments{
\item{vector.1}{The first vector.}

\item{vector.2}{The second vector.}

\item{\dots}{Optional parameters to be passed to low level functions.}
}
\value{
A list with two vectors, arranged according to the input order
}
\description{
This function takes two vectors, combines their factor levels, and return a
list with the same vectors, only with identical and full factor levels. The
order of the factors is first all the factors of vector.1 then the added
factor levels of vector.2. As \code{randomForest} cannot predict if any input
variables of a \code{new.data} contains factor levels that were not included
in the training data, we recommend running this function on each categorical
input variable of the training and \code{new.data} before running the
hierarchical randomForest with \code{\link{RunHRF}}.
}
\examples{
# create two vectors
vec.1 <- as.factor(rep(c("a","b","c"),10))
vec.2 <- as.factor(rep(c("g","a","f","b"),20))

# run the function
new.Vec <- JoinLevels(vec.1,vec.2)

# Re-assign to the original vectors
vec.1 <- new.Vec[[1]]
vec.2 <- new.Vec[[2]]

# note the levels
levels(vec.1)
levels(vec.2)
}
\author{
Yoni Gavish <gavishyoni@gmail.com>
}

