% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Multiplicative_Prop_Votes.R
\name{Multiplicative_Prop_Votes}
\alias{Multiplicative_Prop_Votes}
\title{For each case, the multiplicative proportion of votes}
\usage{
Multiplicative_Prop_Votes(Prop_Vote, Unique_Path, All_Levels = FALSE, ...)
}
\arguments{
\item{Prop_Vote}{Data frame, the proportion of votes that each case received in each local classifer.}
}
\description{
This function takes as input the proporton of votes that each case received
in each local classifer (See: Extract_Votes) and returns the multiplication
of votes, along each path down the class hierarchy. The function prints an
error message for cases whose sum over all terminal nodes is different from 1
(upto 10 digits accuracy).
}
\author{
Yoni Gavish <gavishyoni@gmail.com>
}

