% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Perm_Node_For_Case.R
\name{Perm_Node_For_Case}
\alias{Perm_Node_For_Case}
\title{Randomly select a class from theinbag trees down the hierarchical class structure}
\usage{
Perm_Node_For_Case(Case_Data, Case_ID = 1, Exp_Var = 2:ncol(Case_Data),
  Hie_RF, ...)
}
\description{
STILL UNDER DEVELOPMENT!!!
This function is for a single case. It dependes on individual trees in each
local classifer and as such can only work when \code{Run_HRF} was set to
\code{keep.inbag} = \code{TRUE}.
For a given case, it starts in the tree root and randomly select one tree in
which the case was OOB (Out Of Bag). THepredicted clas of this tree is set as
the new parent class and returned if the class is a terminal node. If the
class is an internal node, the next local classifer down the hierarchy from
the new parent node is identified and the process continues: A random tree in
which the case is OOB is choiosen and its predicted class is set as the new
parent node. This continues untila terminal node is reached. Over multiple
permutation, the proportion of times in which each terminal node is sleceted
will converge to the \code{Mulltiplicative_Prop_Votes}.
}

