% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Plot_Cat_Cat_Var.R
\name{Plot_Cat_Cat_Var}
\alias{Plot_Cat_Cat_Var}
\title{Ploting function for the hierarchical importance}
\usage{
Plot_Cat_Cat_Var(Input_Data, X_Data = 2, Y_Data = 3, Imp_Data = 4,
  Plot_Type = "Tile", Imp_title = colnames(Input_Data)[Imp_Data],
  X_Title = colnames(Input_Data)[X_Data],
  Y_Title = colnames(Input_Data)[Y_Data], low_col = "blue",
  high_col = "red", Geom_Tile_Bor_Col = "white", Pos_Col = "green",
  Zero_col = "white", Neg_Col = "red", ...)
}
\description{
Ploting function for the hierarchical importance
}

