% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Plot_Hie_Tree.R
\name{Plot_Hie_Tree}
\alias{Plot_Hie_Tree}
\title{Plot the Hierarchical class structure}
\usage{
Plot_Hie_Tree(Nodes_Info, Unique_Path, Rect_Width = 0.8, Rect_Hieght = 0.5,
  Text_Size = 7.5, Text_Angle = 0, Split_Text = 6, Int_Col = "white",
  Term_col = "gray80", Root_Col = "gray50",
  Y_Title = "Hierarchical Level\\n", Label_Nodes = TRUE,
  Label_Classifiers = TRUE, Class_ID_Rect_Col = "black",
  Class_ID_Rect_Fill = "gold", Class_ID_Rect_Linetype = "dotted",
  Class_ID_Rect_Hieght_Scale = 0.25, Class_ID_Text_Size_Scale = 0.8,
  Class_ID_Text_Color = "black", ...)
}
\description{
Plot the Hierarchical class structure
}

