% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotImportanceHie.R
\name{PlotImportanceHie}
\alias{PlotImportanceHie}
\title{Plotting function for the hierarchical importance}
\usage{
PlotImportanceHie(input.data, X.data = 2, Y.data = 3, imp.data = 4,
  plot.type = "Tile", imp.title = colnames(input.data)[imp.data],
  X.Title = colnames(input.data)[X.data],
  Y.Title = colnames(input.data)[Y.data], low.col = "blue",
  high.col = "red", geom.tile.bor.col = "white", pos.col = "green",
  zero.col = "white", neg.col = "red", supp.warn = TRUE, ...)
}
\arguments{
\item{input.data}{Data frame, data to be used for plotting.}

\item{X.data}{The column number in \code{input.data} that
  contains the categories for the X axis.}

\item{Y.data}{The column number in \code{input.data} that
  contains the categories for the Y axis.}

\item{imp.data}{The column number in \code{input.data} that
  contains the variable importance values (or any other quantitative vector)
  that will scale the bubble size or tile colour.}

\item{plot.type}{The type of plot to produce. Can be either
  \code{"Tile"} (default) \code{"Bubble"}.}

\item{imp.title}{Character, title for the Importance legend.}

\item{X.Title}{Character, title for the x axis.}

\item{Y.Title}{Character, title for the y axis.}

\item{low.col}{Character, if \code{plot.type} is set to
  \code{"Tile"}, the colour to use for the lowest importance value.}

\item{high.col}{Character, if \code{plot.type} is set to
  \code{"Tile"}, the colour to use for the highest importance value.}

\item{geom.tile.bor.col}{Character, if \code{plot.type} is set to
  \code{"Tile"}, the colour to use for the border of tiles. Set to NA for no
  tile borders.}

\item{pos.col}{Character, if \code{plot.type} is set to
  \code{"Bubble"}, the colour to use for importance values > 0.}

\item{zero.col}{Character, if \code{plot.type} is set to
  \code{"Bubble"}, the colour to use for importance values = 0.}

\item{neg.col}{Character, if \code{plot.type} is set to
  \code{"Bubble"}, the colour to use for importance values < 0.}

\item{supp.warn}{Logical, if \code{TRUE}, warning will not be
  printed. Otherwise, the function may print warnings regarding deprecated
  factor levels.}

\item{\dots}{Optional parameters to be passed to low level functions.}
}
\value{
The function returns a plot. If the function is returned to an
   object, then the plot can be accessed and further edited using $plot on
   the new object (resulting with an object of class \code{ggplot}).
}
\description{
This function takes the 4 column list generated by \code{ImportanceHie} and
plots the importance of each variable in each local classifier as either a
bubble plot or a heat map.
}
\details{
Based on the package \code{\link{ggplot2}}. The generated plot has
  the two categorical values as the X and Y axis. For each combination of X
  and Y, the plot present the variable importance either as a heat map (when
  setting \code{plot.type = "Tile"}) or as a bubble plot (when setting
  \code{plot.type = "Bubble"}).
}
\examples{
set.seed(354)
random.hRF <- RandomHRF(num.term.nodes = 20, tree.depth = 4)
train.data <- random.hRF$train.data

hie.RF.random <- RunHRF(train.data = train.data,
                        case.ID    = "case.ID",
                        hie.levels = c(2:(random.hRF$call$tree.depth + 1)))

Importance.hie.RF <- ImportanceHie(hie.RF     = hie.RF.random,
                                   format.out = c("col.4.out"))

PlotImportanceHie(input.data = Importance.hie.RF,
                  X.data     = 2,
                  Y.data     = 3,
                  imp.data   = 4,
                  plot.type  = "Tile",
                  supp.warn  = FALSE)

# Tile, black and white
# label with the classifier.ID instead of the parent node name.
impor.plot.tile <- PlotImportanceHie(input.data        = Importance.hie.RF,
                                     X.data            = 1,
                                     Y.data            = 3,
                                     imp.data          = 4,
                                     plot.type         = "Tile",
                                     low.col           = "white" ,
                                     high.col          = "black",
                                     geom.tile.bor.col = "gray20",
                                     supp.warn         = TRUE)

# further editing of the plot with ggplot2
impor.plot.tile.2 <- impor.plot.tile$plot +
                     theme(axis.title.y = element_text(size   = 20,
                                                       colour = "red"))
impor.plot.tile.2 <- impor.plot.tile.2 +
   ggtitle("The mean decrease in accuracy of each
   explanatory variable in each local classifier")
impor.plot.tile.2

#Bubble
PlotImportanceHie(input.data = Importance.hie.RF,
                  X.data     = 2,
                  Y.data     = 3,
                  imp.data   = 4,
                  plot.type  = "Bubble")

#Bubble, black and white
PlotImportanceHie(input.data = Importance.hie.RF,
                  X.data     = 2,
                  Y.data     = 3,
                  imp.data   = 4,
                  plot.type  = "Bubble",
                  pos.col    = "black",
                  zero.col   = "gray50",
                  neg.col    = "white")
}
\author{
Yoni Gavish <gavishyoni@gmail.com>
}
\seealso{
\code{\link{ImportanceHie}} for extracting importance values from an
  object of class \code{"HRF"}.
}

