% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Predict_New.R
\name{Predict_New}
\alias{Predict_New}
\title{predict function for New_Data using an existing  object of class \code{"Hier.Random.Forest"}}
\usage{
Predict_New(Hie_RF, New_Data, New_Data_Case_ID = 1,
  New_Data_Exp_Var = c(2:ncol(New_Data)),
  Crisp_Rule = c("Multiplicative_Majority", "Multiplicative_Permutation",
  "Setpwise_Majority"), Perm_Num = 500, Div_Logical = TRUE,
  Div_Print = 25)
}
\description{
predict function for New_Data using an existing  object of class \code{"Hier.Random.Forest"}
}

