% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Random_HRF_Data.R
\name{Random_HRF_Data}
\alias{Random_HRF_Data}
\title{Create a random class hierarchy tree and random training and New data thtat fits it}
\usage{
Random_HRF_Data(Num_Term_Nodes = 25, Tree_Depth = 5,
  Cells_per_Cat_Train = 15, Num_Num_Var = 10, Num_Cat_Var = 10,
  Case_ID = "Case_ID", cells_per_cat_new = 25, ...)
}
\description{
This funtion uses the \code{random.hierarchical.data} function of the package
\emph{"treemap"} to create a tree hierarchy with a given number of terminal
nodes and with a given number of hierarchical levels. Next, random quantitative and
categorical explanatory variables are added to a user-defined number of
training and new_data cases of each terminal class.
}
\note{
Currently, the function can only return hierarchies that end at the
  lowest level. Future versions may allow trimmed hierarchy trees.
}

