% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RandomHRF.R
\name{RandomHRF}
\alias{RandomHRF}
\title{Create random class hierarchy with random training and \code{new.data} cases.}
\usage{
RandomHRF(num.term.nodes = 25, tree.depth = 5, cases.t.node.train = 15,
  cases.t.node.new = 25, new.data.observed = FALSE, numer.exp.var = 10,
  categ.exp.var = 10, case.ID = "case.ID", ...)
}
\arguments{
\item{num.term.nodes}{Integer, the number of terminal nodes at the
  lowest level.}

\item{tree.depth}{Integer, the number of levels in the class
  hierarchy.}

\item{cases.t.node.train}{Integer, the number of training cases to
  create for each terminal node in the training data.}

\item{cases.t.node.new}{Integer, the number of training cases to
  create for each terminal node in \code{new.data}.}

\item{new.data.observed}{Logical, if \code{FALSE} (default)- the
  \code{new.data} will not contain an observed class hierarchy for each case.
  If \code{TRUE} an observed class hierarchy will be added to each case of
  \code{new.data}.}

\item{numer.exp.var}{Integer, number of numeric explanatory
  variables.}

\item{categ.exp.var}{Integer, number of categorical explanatory
  variables.}

\item{case.ID}{Character, the name of the column that will
  contain the \code{case.IDs}.}

\item{\dots}{Optional parameters to be passed to low level functions.}
}
\value{
A list consisting of:
\item{train.data}{Data frame with the training dataset}
\item{new.data}{Data frame with the \code{new.data} dataset}
\item{call}{The call to \code{RandomHRF}}
}
\description{
This function uses the \code{\link{random.hierarchical.data}} function of the
package \code{"treemap"} to create a tree hierarchy with a given number of
terminal nodes and with a given number of hierarchical levels. Next, random
quantitative and categorical explanatory variables are added to a
user-defined number of training and \code{new.data} cases of each terminal
node.
}
\details{
The numeric explanatory variables are based on sampling with
  replacement from a vector of integers starting with 1 and ending with a
  random number between 1 and 50. The categorical explanatory variables are
  based on sampling with replacement of the first 10 letters. Future version
  may allow greater control on the creation of explanatory variables. \cr
  Setting \code{cases.t.node.train = 0} or \code{cases.t.node.new = 0} will
  result with \code{NA} for \code{train.data} or \code{new.data},
  respectively.
}
\note{
Currently, the function can only return hierarchies for which all
  terminal nodes occur at the lowest level. Future versions may allow trimmed
  hierarchy trees.
}
\examples{
set.seed(354)
random.hRF <- RandomHRF(num.term.nodes = 20,
                        tree.depth     = 4)

train.data <- random.hRF$train.data
new.data   <- random.hRF$new.data
random.hRF$call$tree.depth
#  example with new.data.observed =TRUE
set.seed(354)
random.hRF <- RandomHRF(num.term.nodes = 15,
                        tree.depth     = 6,
                        new.data.observed = TRUE)
train.data <- random.hRF$train.data
new.data   <- random.hRF$new.data
}
\author{
Yoni Gavish <gavishyoni@gmail.com>
}

