% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Run_HRF_As_Flat.R
\name{Run_HRF_As_Flat}
\alias{Run_HRF_As_Flat}
\title{Run the same training data as flat randomForest classifcation}
\usage{
Run_HRF_As_Flat(Hie_RF, mtry = Hie_RF$call$mtry, ntree = Hie_RF$call$ntree,
  importance = Hie_RF$call$importance, proximity = Hie_RF$call$proximity,
  keep.forest = Hie_RF$call$keep.forest,
  keep.inbag = Hie_RF$call$keep.inbag, ...)
}
\description{
Run the same training data as flat randomForest classifcation
}

