% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tuneRF_2.R
\name{tuneRF_2}
\alias{tuneRF_2}
\title{the tuneRF function of randomForest after correcting for error relating to errorOld=0}
\usage{
tuneRF_2(x, y, mtryStart = if (is.factor(y)) floor(sqrt(ncol(x))) else
  floor(ncol(x)/3), ntreeTry = 50, stepFactor = 2, improve = 0.05,
  trace = TRUE, plot = FALSE, doBest = FALSE, ...)
}
\description{
the tuneRF function of randomForest after correcting for error relating to errorOld=0
}

